Rename to EveMapperEntity to EveMapperService
