+++ src/WHMapper/Program.cs
@@ -212,7 +212,7 @@ namespace WHMapper
             #endregion
 
             #region WH HELPER
-            builder.Services.AddScoped<IEveMapperEntity, EveMapperEntity>();
+            builder.Services.AddScoped<IEveMapperService, EveMapperService>();
             builder.Services.AddScoped<IEveMapperAccessHelper, EveMapperAccessHelper>();
             builder.Services.AddScoped<IEveMapperTracker, EveMapperTracker>();
             builder.Services.AddScoped<IEveMapperSearch, EveMapperSearch>();
